# irmin-pack: remove unecessary calls to Index.find