--- a/src/irmin-pack/indexing_strategy.ml
+++ b/src/irmin-pack/indexing_strategy.ml
@@ -50,3 +50,4 @@ let minimal_with_contents : t =
   | Dangling_parent_commit -> assert false
 
 let default = always
+let is_minimal x = x == minimal