--- a/src/irmin-pack/indexing_strategy.mli
+++ b/src/irmin-pack/indexing_strategy.mli
@@ -42,3 +42,6 @@ val default : t
 (** [default] is the indexing strategy used by [irmin-pack] instances that do
     not explicitly set an indexing strategy in {!Irmin_pack.config}. Currently
     set to {!always}. *)
+
+val is_minimal : t -> bool
+(** [is_minimal t] is true if [t] is {!minimal}. *)