In minimal indexing mode we don't need to check the index file too often

This removes a few calls to index in the benchmarks:

before:
```
+009us [DEBUG] irmin-pack.unix [pack] batch start
+012us [DEBUG] irmin-pack.unix index CoVkT31UeHdS8o4MMQtzR5jmdxZ5PHo9C1P7LLvn9sSvYub5tB3q
+040us [DEBUG] index [root] find CoVkT31UeHdS8o4MMQtzR5jmdxZ5PHo9C1P7LLvn9sSvYub5tB3q
+023us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoVkT31UeHdS8o4MMQtzR5jmdxZ5PHo9C1P7LLvn9sSvYub5tB3q",0,41]}
+028us [DEBUG] irmin.pack save depth:0
+010us [DEBUG] irmin-pack.unix index CoUr3AHPmaXCuYL2ps3bcKXwex2QmngoT9Z8LhhMWdLw8RTWZbGc
+010us [DEBUG] index [root] find CoUr3AHPmaXCuYL2ps3bcKXwex2QmngoT9Z8LhhMWdLw8RTWZbGc
+011us [DEBUG] irmin-pack.unix [dict] index "version"
+010us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoUr3AHPmaXCuYL2ps3bcKXwex2QmngoT9Z8LhhMWdLw8RTWZbGc",41,46]}
+018us [DEBUG] irmin-pack.unix index CoWMouGpUV1DJEMBUvZRToG6e2H2JK7FbW2xxMMyECWfvJZ4pNfZ
+009us [DEBUG] index [root] find CoWMouGpUV1DJEMBUvZRToG6e2H2JK7FbW2xxMMyECWfvJZ4pNfZ
+011us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoWMouGpUV1DJEMBUvZRToG6e2H2JK7FbW2xxMMyECWfvJZ4pNfZ",87,66]}
+015us [DEBUG] irmin-pack.unix index CoVa898LzEftSsbRB5gHhNtRqUwsVDoS1gw6fs4LtNQQDgzjXzfh
+010us [DEBUG] index [root] find CoVa898LzEftSsbRB5gHhNtRqUwsVDoS1gw6fs4LtNQQDgzjXzfh
+010us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoVa898LzEftSsbRB5gHhNtRqUwsVDoS1gw6fs4LtNQQDgzjXzfh",153,35]}
+016us [DEBUG] irmin.pack save depth:0
+008us [DEBUG] irmin-pack.unix index CoVmWcmHnA6efWzV8PS7fdVSX2EaJsLCU5uj231DpzizqSv8b271
+010us [DEBUG] index [root] find CoVmWcmHnA6efWzV8PS7fdVSX2EaJsLCU5uj231DpzizqSv8b271
+010us [DEBUG] irmin-pack.unix [dict] index "test_chain"
+006us [DEBUG] irmin-pack.unix [dict] index "protocol"
+005us [DEBUG] irmin-pack.unix [dict] index "data"
+007us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoVmWcmHnA6efWzV8PS7fdVSX2EaJsLCU5uj231DpzizqSv8b271",188,66]}
+017us [DEBUG] irmin-pack.unix index CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj
+010us [DEBUG] index [root] find CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj
+015us [DEBUG] index [root] replace CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj [254,66,"D"]
+019us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj",254,66]}
+014us [INFO] irmin-pack.unix [pack] batch completed in 0.35ms
```

after:

```
+009us [DEBUG] irmin-pack.unix [pack] batch start
+018us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoVkT31UeHdS8o4MMQtzR5jmdxZ5PHo9C1P7LLvn9sSvYub5tB3q",0,41]}
+052us [DEBUG] irmin.pack save depth:0
+012us [DEBUG] irmin-pack.unix [dict] index "version"
+011us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoUr3AHPmaXCuYL2ps3bcKXwex2QmngoT9Z8LhhMWdLw8RTWZbGc",41,46]}
+019us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoWMouGpUV1DJEMBUvZRToG6e2H2JK7FbW2xxMMyECWfvJZ4pNfZ",87,66]}
+016us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoVa898LzEftSsbRB5gHhNtRqUwsVDoS1gw6fs4LtNQQDgzjXzfh",153,35]}
+016us [DEBUG] irmin.pack save depth:0
+008us [DEBUG] irmin-pack.unix [dict] index "test_chain"
+006us [DEBUG] irmin-pack.unix [dict] index "protocol"
+005us [DEBUG] irmin-pack.unix [dict] index "data"
+006us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoVmWcmHnA6efWzV8PS7fdVSX2EaJsLCU5uj231DpzizqSv8b271",188,66]}
+018us [DEBUG] irmin-pack.unix index CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj
+011us [DEBUG] index [root] find CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj
+018us [DEBUG] index [root] replace CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj [254,66,"D"]
+017us [DEBUG] irmin-pack.unix [pack] append {"Direct":["CoV8SQumiVU9saiu3FVNeDNewJaJH8yWdsGF3WLdsRr2P9S7MzCj",254,66]}
+014us [INFO] irmin-pack.unix [pack] batch completed in 0.24ms
```