@samoht 2022-07-21
The IO read stats seem better with this on micro-benchmarks - need to see how this scales?

<img width="1235" alt="Screenshot 2022-07-21 at 09 26 42" src="https://user-images.githubusercontent.com/103693/180155034-2ef0564b-30f0-47bb-b79e-df1957e471c4.png">