@samoht 2022-07-21
@icristescu thanks, the resulting patch is much more smaller now :-)